requirements.txt
discord.py
random
PyNaCl
youtube-dl